mqtt_tester/pyproject.toml
[project]
name = "mqtt-tester"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["dash>=3.2.0", "matplotlib>=3.10.7", "paho-mqtt>=2.1.0"]
